setting up styling for landing page and homepage

+++ frontend/src/pages/AccountPage/index.js
@@ -3,6 +3,7 @@ import { useNavigate } from "react-router-dom"
 import { Link } from "react-router-dom"
 import Button from 'react-bootstrap/Button';
 import React, { useEffect, useState } from "react";
+import './accountPage.css'
 
 
 const AccountPage = (props) => {
@@ -18,41 +19,40 @@ const AccountPage = (props) => {
     }, [props.currentUser]);
 
     const handleDelete = () => {
-        deleteUserAccount(userData._id);
-        localStorage.clear();
-        props.setIsLoggedIn(false);
-        navigate('/');
+        if (window.confirm("Are you sure you want to delete your account? This action can't be undone!")) {
+            deleteUserAccount(userData._id);
+            localStorage.clear();
+            props.setIsLoggedIn(false);
+            navigate('/');
+        }
     }
 
     return(
         <>
         <div className="account-page">
             <div>
-                <div className="coloumn sm-6 account-greeting">
-                    <div className="account-info">
-                        {/* <h1>Hello, {userData.username ? userData.username : 'User'}</h1> */}
-                        <p>HODL or die trying</p>
-                        <h4>Username: {userData.username}</h4>
-                        <p> Want to create a new comment? If so, click on the button below!</p>
-                        <Button href="/newcomment" id="create-comment">Create Listing</Button>
-                        <p>Permantly delete your account, active listings, and all data associated with it.</p>
-                        <p id="warning"> Warning: this action can't be undone!</p> 
-                        <Button id="delete-account" onClick={handleDelete}>Delete Account</Button>
+                <div className="account-info" >
+                    <div className="coloumn sm-6 account-greeting">
+                        <h2>Hello {userData.username ? userData.username : 'User'}, Thank You for using Cryptosis!</h2>
+                        <h4>Your Username: {userData.username}</h4>
+                        <p> Want to share your cryptocurrency thoughts on the discussion board? If so, click on the button below!</p>
+                        <Button href="/newcomment" id="create-comment">Create a Comment</Button>
+                        <br />
+                        <br />
+                        <Button href="/" variant="secondary" id="go-home">View Discussion Board</Button>
                     </div>
                 </div>
             </div>
         </div>
         <div className="active-listings">
-            <h2>Your Active Listings:</h2>
+            <h2>Your Active Comments:</h2>
             <div className="row">
                 {userCommentData.map((comment, i) => {
-                    if (!userData) return null;
-                    return(
+                return(
                     <div key={i} className="card" style={{width: '18rem'}}>
-                        <img className="card-img-top" src={comment.image} alt={comment.title}/>
-                        <div className="card-body">
-                            <h5 className="card-title">{comment.title}</h5>
-                            <p className="card-text">{comment.description}</p>
+                        <div>
+                            <h5 >{comment.title}</h5>
+                            <p >{comment.description}</p>
                             <div>
                                 <Link to={"/commentedit/" + comment._id}>Edit</Link>
                             </div>
@@ -60,6 +60,11 @@ const AccountPage = (props) => {
                     </div>
                     )
                 })}
+            </div>
+            <div className="deleteAccountDiv">
+            <p>Permantly delete your account, active listings, and all data associated with it.</p>
+                        <p id="warning"> Warning: this action can't be undone!</p> 
+                        <Button id="delete-account" onClick={handleDelete}>Delete Account</Button>
             </div>
         </div>
         </>
